+++ index.js
@@ -49,6 +49,7 @@ function compression(options) {
   }
 
   return function compression(req, res, next){
+    var ended = false
     var length
     var listeners = []
     var write = res.write
@@ -56,17 +57,16 @@ function compression(options) {
     var end = res.end
     var stream
 
-    // see #8
-    req.on('close', function(){
-      res.write = res.end = noop
-    });
-
     // flush is noop by default
     res.flush = noop;
 
     // proxy
 
     res.write = function(chunk, encoding){
+      if (ended) {
+        return false
+      }
+
       if (!this._header) {
         this._implicitHeader()
       }
@@ -77,6 +77,10 @@ function compression(options) {
     };
 
     res.end = function(chunk, encoding){
+      if (ended) {
+        return false
+      }
+
       if (!this._header) {
         // estimate the length
         if (!this.getHeader('Content-Length')) {
@@ -89,6 +93,9 @@ function compression(options) {
       if (!stream) {
         return end.call(this, chunk, encoding)
       }
+
+      // mark ended
+      ended = true
 
       // write Buffer for Node.js 0.8
       return chunk