+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Fix return value from `.end` and `.write` after end
   * Improve detection of zero-length body without `Content-Length`
   * deps: accepts@~1.2.9
     - deps: mime-types@~2.1.1